refactor abstract away config class
