+++ src/zeroband/infer.py
@@ -22,7 +22,7 @@ from vllm import LLM, SamplingParams, TokensPrompt
 from huggingface_hub import snapshot_download
 
 from zeroband.utils.config import extract_toml_paths, to_kebab_case
-from zeroband.inference.config import Config as InferenceConfig, set_toml_paths
+from zeroband.inference.config import Config as InferenceConfig
 from zeroband.inference.parquet import get_parquet_table
 from zeroband.inference.pipeline import all_reduce, patch_model_load, setup_comm, setup_hooks
 from zeroband.inference.rewards import compute_vllm_rewards
@@ -443,7 +443,7 @@ if __name__ == "__main__":
 
     # Extract toml file paths from CLI arguments
     toml_paths, cli_args = extract_toml_paths(sys.argv[1:])
-    set_toml_paths(toml_paths)
+    InferenceConfig.set_global_toml_files(toml_paths)
 
     config = InferenceConfig(_cli_parse_args=to_kebab_case(cli_args))
